+++ src/utils/helpers/browser/__spec__.js
@@ -1,10 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { mount } from 'enzyme';
 import Form from './../../../components/form';
 import Browser from './browser.js';
 
 describe('Browser', () => {
   let _window;
+  const elem = mount(
+    <div className='foo bar baz' />
+  );
 
   beforeEach(() => {
     _window = {
@@ -14,7 +18,7 @@ describe('Browser', () => {
 
   describe('redirectTo', () => {
     describe('when url is passed', () => {
-      let urlsample = 'http://bla';
+      const urlsample = 'http://bla';
 
       it('redirects to url', () => {
         spyOn(Browser, 'getWindow').and.returnValue(_window);
@@ -45,7 +49,7 @@ describe('Browser', () => {
 
   describe('reload', () => {
     it('calls the windows location relaod method', () => {
-      let spy = jasmine.createSpy('reload');
+      const spy = jasmine.createSpy('reload');
       _window = { location: { reload: spy } };
 
       spyOn(Browser, 'getWindow').and.returnValue(_window);
@@ -92,7 +96,7 @@ describe('Browser', () => {
       describe('options', () => {
         describe('when expires is passed', () => {
           it('adds expires date to the cookie', () => {
-            let date = new Date();
+            const date = new Date();
             date.setDate(date.getDate() + 1);
             Browser.setCookie('foo', 'bar', { expires: date });
             // Cannot check expiration date from document api
@@ -100,7 +104,7 @@ describe('Browser', () => {
           });
 
           it('does not add a cookie when expiration is in the passed', () => {
-            let date = new Date();
+            const date = new Date();
             date.setDate(date.getDate() - 1);
             Browser.setCookie('foo', 'bar', { expires: date });
             // Cannot check expiration date from document api
@@ -184,8 +188,14 @@ describe('Browser', () => {
     it('renders a hidden input for each data member', () => {
       spyOn(React, 'createElement').and.callThrough();
       Browser.postToNewWindow(url, data);
-      expect(React.createElement).toHaveBeenCalledWith('input', { type: 'hidden', key: key1, name: key1, value: value1 });
-      expect(React.createElement).toHaveBeenCalledWith('input', { type: 'hidden', key: key2, name: key2, value: value2 });
+      expect(React.createElement).toHaveBeenCalledWith(
+        'input',
+        { type: 'hidden', key: key1, name: key1, value: value1 }
+      );
+      expect(React.createElement).toHaveBeenCalledWith(
+        'input',
+        { type: 'hidden', key: key2, name: key2, value: value2 }
+      );
     });
 
     it('submits the rendered form', () => {
@@ -205,11 +215,11 @@ describe('Browser', () => {
     describe('when target option is passed', () => {
       it('sets the form target', () => {
         spyOn(React, 'createElement').and.callThrough();
-        let target = 'some_window';
+        const target = 'some_window';
 
         Browser.postToNewWindow(url, data, target);
         expect(React.createElement).toHaveBeenCalledWith(Form, {
-          action: url, method: 'post', target: target, save: false, cancel: false
+          action: url, method: 'post', target, save: false, cancel: false
         }, jasmine.anything());
       });
     });
@@ -226,7 +236,7 @@ describe('Browser', () => {
 });
 
 function addPostFormDiv() {
-  const div = document.createElement("div");
+  const div = document.createElement('div');
   div.setAttribute('id', 'carbonPostFormContainer');
   document.body.appendChild(div);
 }